requirements.txt
aiogram
pydantic
python-dotenv
aiosqlite
